(restructure) - hoist more component logic (#3316)
* introduce skip_children flag

* hoist context logic and reduce bailout logic

* optimize mount route
