+++ src/constants.js
@@ -45,6 +45,8 @@ export const MODE_SVG = 1 << 12;
 export const FORCE_UPDATE = 1 << 13;
 /** Signifies that a node needs to be updated */
 export const DIRTY_BIT = 1 << 14;
+/** Signals the component can skip children due to a non-update */
+export const SKIP_CHILDREN = 1 << 15;
 
 /** Reset all mode flags */
 export const RESET_MODE = ~(
@@ -53,7 +55,8 @@ export const RESET_MODE = ~(
 	MODE_SUSPENDED |
 	MODE_ERRORED |
 	MODE_RERENDERING_ERROR |
-	FORCE_UPDATE
+	FORCE_UPDATE |
+	SKIP_CHILDREN
 );
 
 /** Modes a child internal inherits from their parent */
