+++ src/diff/component.js
@@ -1,17 +1,15 @@
-import { Fragment } from '../create-element';
 import options from '../options';
-import { mountChildren } from './mount';
-import { diffChildren, reorderChildren } from './children';
-import { DIRTY_BIT, FORCE_UPDATE } from '../constants';
+import { DIRTY_BIT, FORCE_UPDATE, SKIP_CHILDREN } from '../constants';
 import { addCommitCallback } from './commit';
-import { getParentContext } from '../tree';
 
 export function renderFunctionComponent(
 	parentDom,
 	newVNode,
 	internal,
 	commitQueue,
-	startDom
+	startDom,
+	context,
+	componentContext
 ) {
 	/** @type {import('../internal').Component} */
 	let c;
@@ -23,18 +21,6 @@ export function renderFunctionComponent(
 	// @TODO split update + mount?
 	let newProps = newVNode ? newVNode.props : internal.props;
 
-	const context = getParentContext(internal);
-
-	// Necessary for createContext api. Setting this property will pass
-	// the context value as `this.context` just for this component.
-	tmp = type.contextType;
-	let provider = tmp && context[tmp._id];
-	let componentContext = tmp
-		? provider
-			? provider.props.value
-			: tmp._defaultValue
-		: context;
-
 	if (internal && internal._component) {
 		c = internal._component;
 	} else {
@@ -44,18 +30,13 @@ export function renderFunctionComponent(
 			forceUpdate: internal.rerender.bind(null, internal)
 		};
 
-		if (provider) provider._subs.add(internal);
-
 		internal.flags |= DIRTY_BIT;
 	}
 
 	if (newVNode && newVNode._vnodeId === internal._vnodeId) {
-		internal.props = c.props = newProps;
-		c._internal = internal;
-		// TODO: Returning undefined here (i.e. return;) passes all tests. That seems
-		// like a bug. Should validate that we have test coverage for sCU that
-		// returns Fragments with multiple DOM children
-		return reorderChildren(internal, startDom, parentDom);
+		c.props = newProps;
+		internal.flags |= SKIP_CHILDREN;
+		return;
 	}
 
 	c.context = componentContext;
@@ -72,31 +53,7 @@ export function renderFunctionComponent(
 		internal._context = Object.assign({}, context, c.getChildContext());
 	}
 
-	let isTopLevelFragment =
-		tmp != null && tmp.type === Fragment && tmp.key == null;
-	let renderResult = isTopLevelFragment ? tmp.props.children : tmp;
-
-	let nextDomSibling;
-
-	if (internal._children == null) {
-		nextDomSibling = mountChildren(
-			parentDom,
-			Array.isArray(renderResult) ? renderResult : [renderResult],
-			internal,
-			commitQueue,
-			startDom
-		);
-	} else {
-		nextDomSibling = diffChildren(
-			parentDom,
-			Array.isArray(renderResult) ? renderResult : [renderResult],
-			internal,
-			commitQueue,
-			startDom
-		);
-	}
-
-	return nextDomSibling;
+	return tmp;
 }
 
 /**
@@ -114,7 +71,9 @@ export function renderClassComponent(
 	newVNode,
 	internal,
 	commitQueue,
-	startDom
+	startDom,
+	context,
+	componentContext
 ) {
 	/** @type {import('../internal').Component} */
 	let c;
@@ -126,25 +85,11 @@ export function renderClassComponent(
 	// @TODO split update + mount?
 	let newProps = newVNode ? newVNode.props : internal.props;
 
-	const context = getParentContext(internal);
-
-	// Necessary for createContext api. Setting this property will pass
-	// the context value as `this.context` just for this component.
-	tmp = type.contextType;
-	let provider = tmp && context[tmp._id];
-	let componentContext = tmp
-		? provider
-			? provider.props.value
-			: tmp._defaultValue
-		: context;
-
 	if (internal && internal._component) {
 		c = internal._component;
 	} else {
 		// @ts-ignore The check above verifies that newType is suppose to be constructed
 		internal._component = c = new type(newProps, componentContext); // eslint-disable-line new-cap
-
-		if (provider) provider._subs.add(internal);
 
 		if (!c.state) c.state = {};
 		isNew = true;
@@ -195,19 +140,10 @@ export function renderClassComponent(
 					false) ||
 			(newVNode && newVNode._vnodeId === internal._vnodeId)
 		) {
-			internal.props = c.props = newProps;
+			c.props = newProps;
 			c.state = c._nextState;
-			// More info about this here: https://gist.github.com/JoviDeCroock/bec5f2ce93544d2e6070ef8e0036e4e8
-			if (newVNode && newVNode._vnodeId !== internal._vnodeId) {
-				internal.flags &= ~DIRTY_BIT;
-			}
-
-			c._internal = internal;
-
-			// TODO: Returning undefined here (i.e. return;) passes all tests. That seems
-			// like a bug. Should validate that we have test coverage for sCU that
-			// returns Fragments with multiple DOM children
-			return reorderChildren(internal, startDom, parentDom);
+			internal.flags |= SKIP_CHILDREN;
+			return;
 		}
 
 		if (c.componentWillUpdate != null) {
@@ -246,29 +182,5 @@ export function renderClassComponent(
 		}
 	}
 
-	let isTopLevelFragment =
-		tmp != null && tmp.type === Fragment && tmp.key == null;
-	let renderResult = isTopLevelFragment ? tmp.props.children : tmp;
-
-	let nextDomSibling;
-
-	if (internal._children == null) {
-		nextDomSibling = mountChildren(
-			parentDom,
-			Array.isArray(renderResult) ? renderResult : [renderResult],
-			internal,
-			commitQueue,
-			startDom
-		);
-	} else {
-		nextDomSibling = diffChildren(
-			parentDom,
-			Array.isArray(renderResult) ? renderResult : [renderResult],
-			internal,
-			commitQueue,
-			startDom
-		);
-	}
-
-	return nextDomSibling;
+	return tmp;
 }
